backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
joblib>=1.3.0
Pillow>=10.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
speechrecognition>=3.10.0
pyttsx3>=2.90
gtts>=2.3.0
pyaudio>=0.2.11
requests>=2.31.0
aiofiles>=23.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
APScheduler>=3.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
googletrans>=4.0.0